Add state machine-based tests for MinHeap
At the moment this might seem circular because of the use of heappop and heappush
when they're also used in the module, but the MinHeap ADT's API is
a little different from the standard heappop and heappush methods. We
test the API differences, in addition to this serving as a regression
test.